Fix compiler warning (#5318)
* fixed compiler maybe-uninitialized warning

* fixed: using delete to release object created by new

* fixed compiler warning: ISO C++ forbids converting a string constant to 'char*' and using ZPP api

* fixed compiler warning: 'recv_size' may be used uninitialized and comparison of integer expressions of different signedness
# Conflicts:
#	ext-src/swoole_postgresql_coro.cc

+++ ext-src/swoole_client_coro.cc
@@ -505,7 +505,7 @@ static PHP_METHOD(swoole_client_coro, peek) {
 
     CLIENT_CORO_GET_SOCKET_SAFE(cli);
 
-    buf = (char *) emalloc(buf_len + 1);
+    buf = (char *) emalloc((size_t)buf_len + 1);
     ret = cli->peek(buf, buf_len);
     if (ret < 0) {
         php_swoole_socket_set_error_properties(ZEND_THIS, cli);
